1062 - Disable remove and create unavailability if sheet does not attend event

+++ src/Application/Command/Unavailability/RemoveHandler.php
@@ -10,6 +10,7 @@
 
 namespace Proximum\Vimeet\Application\Command\Unavailability;
 
+use Proximum\Vimeet\Application\Exception\Unavailability\CanNotDeleteUnavailabilityException;
 use Proximum\Vimeet\Domain\Repository\UnavailabilityRepositoryInterface;
 
 class RemoveHandler
@@ -31,9 +32,15 @@ class RemoveHandler
 
     /**
      * @param Remove $remove
+     *
+     * @throws CanNotDeleteUnavailabilityException
      */
     public function handle(Remove $remove)
     {
+        if (!$remove->unavailability->getParticipant()->getSheet()->attend()) {
+            throw new CanNotDeleteUnavailabilityException('The sheet does not attend the event, therefore the unavailability can not be remove');
+        }
+
         $this->unavailabilityRepository->remove($remove->unavailability);
     }
 }
